Rename .neon file

+++ stats/project_analysis/analyzer.sh
@@ -7,19 +7,24 @@ cd /var/lib/drupalci/workspace/drupal-checkouts/drupal$5
 COMPOSER_CACHE_DIR=/tmp/cache$5 composer --no-interaction --no-progress require drupal/$2 $3 2>> /var/lib/drupalci/workspace/phpstan-results/$1.$3.phpstan_stderr
 php -d memory_limit=2G -d sys_temp_dir=/var/lib/drupalci/workspace/drupal-checkouts/drupal$5 ./vendor/bin/phpstan analyse --no-progress --error-format checkstyle -c ./phpstan.neon  ./${4#project_}s/contrib/$2 > /var/lib/drupalci/workspace/phpstan-results/$1.$3.phpstan_results.xml 2>> /var/lib/drupalci/workspace/phpstan-results/$1.$3.phpstan_stderr
 
-# Remove phpstan.neon because it is not needed for rector and causes some modules to fail.
-rm /var/lib/drupalci/workspace/drupal-checkouts/drupal$5/phpstan.neon
 # Only run rector if we have some file messages in the XML.
+if grep -q '<file name' /var/lib/drupalci/workspace/phpstan-results/$1.$3.phpstan_results.xml; then
+  # Rename phpstan.neon because it is not needed for rector and causes some modules to fail.
+  mv phpstan.neon phpstan.neon.hide
   # Create a git commit for the current state of the project
-cd ${4#project_}s/contrib/$2
-git init
-git add .;git commit -q -m "git project before rector"
-cd /var/lib/drupalci/workspace/drupal-checkouts/drupal$5
-php -d memory_limit=2G -d sys_temp_dir=/var/lib/drupalci/workspace/drupal-checkouts/drupal$5 ./vendor/bin/rector process --verbose ${4#project_}s/contrib/$2 &>  /var/lib/drupalci/workspace/phpstan-results/$1.$3.rector_out
-cd ${4#project_}s/contrib/$2
-git diff > /var/lib/drupalci/workspace/phpstan-results/$1.$3.rector.patch
-# Delete the file if it is empty.
-find /var/lib/drupalci/workspace/phpstan-results/$1.$3.rector.patch -size  0 -print -delete
+  cd ${4#project_}s/contrib/$2
+  git init
+  git add .;git commit -q -m "git project before rector"
+  cd /var/lib/drupalci/workspace/drupal-checkouts/drupal$5
+  php -d memory_limit=2G -d sys_temp_dir=/var/lib/drupalci/workspace/drupal-checkouts/drupal$5 ./vendor/bin/rector process --verbose ${4#project_}s/contrib/$2 &>  /var/lib/drupalci/workspace/phpstan-results/$1.$3.rector_out
+  cd ${4#project_}s/contrib/$2
+  git diff > /var/lib/drupalci/workspace/phpstan-results/$1.$3.rector.patch
+  # Delete the file if it is empty.
+  find /var/lib/drupalci/workspace/phpstan-results/$1.$3.rector.patch -size  0 -print -delete
+  # Restore phpstan.neon
+  cd /var/lib/drupalci/workspace/drupal-checkouts/drupal$5
+  mv phpstan.neon.hide phpstan.neon
+fi
 
 git reset --hard HEAD
 git clean -ffd
